requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
docker-pycreds==0.4.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
lightning-utilities==0.10.0
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
packaging==23.2
Pillow==10.1.0
protobuf==4.25.1 # 3.19.5
psutil==5.9.6
pyparsing==3.1.1
pytorch-lightning
PyYAML==6.0.1
rdkit==2023.9.2
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.2.0
# torch==2.1.0
# torch_geometric==2.4.0
# torchaudio==2.1.0
# torchmetrics==1.2.1
# torchvision==0.16.0
tqdm==4.66.1
triton
typing_extensions==4.8.0
urllib3==2.0.7
wandb==0.16.1
yarl==1.9.4
